Add files via upload
Deliver message by voice is the most important, effective and common method of
exchange information for mankind. Language is human specific features and human
voice is commonly used tool which is also the important way to pass information to
each other. The voice has large information capacity. So we can use modern method to
study and to use voice processing technology . We have created an application that
combines different types of voice processing modules, so that people can easily
transmit, access, change and apply the voice. The objective of the project is to design
an application in MATLAB platform by designing GUI to process voice for different
purposes. This voice processing application demo project is capable either recording a
voice or loading a *.wav file. Once the demo receives the file it provides the option of
playing it or saving it for later use. The modules are also cascaded in the same manner
as they are presented on the GUI. In order to use the GUI of the project we have four
modules which are audio effects, de-noising, speech encryption and speech recognition
. The first module, needs to select between two sub-modules which are music effect
and speech changer after this selection ,user have to determine the type of sound effect
depend on need .The second module, needs to select between two sub-modules which
are low-pass filter and noise suppression demo which need to control of some
parameters to reduce the noise. The third module, request from the user to enter the
recording's seconds to encrypt the recorded sound .The last module, request from the
user to enter the passcode by user's voice one by one and compare it with actual
passcode . For better understanding, the output and input of signals can be plotted on
separate figures. It also provides the two most common ways of plotting signals which
are the time domain, frequency domain. During this project, some features are added, for
example, we have added a feature to handle a set of audio processing methods into one
application and some functions are improved in this sense the thesis achieved its aim.